separate options from config
